Publish the EFS name for an aws region
I could not use aws ec2 describe-instances to get the EFS value from the tags
becasue I need credentials and all the secrets are on the EFS

I decided the filesystem name for each region is not a secret.
The EFS can only be accessed by an instance within the region.

The github pages site is highly available but will become a prerequisite
for mounting the EFS.  All this becasue I already have more than 16 scripts
and don't want than many for each region where I will operate.