Add Create locate database for nfs mounted efs
Add Create locate database for nfs mounted efs section to aws-efs page.

Remove script that was here as part of an IPv6 workaround.

The aws-efs-www-git-clone.bash script was in the aws-efs directory on GitHub
pages as part of the early stages of IPv6 workarounds before I started putting
the public repos on GitLab as well as GitHub.

The aws repo can now be installed early on instances without a public IPv4
address and the aws-efs-www-git-clone.bash script and other programs the
Initialization uses are avaibable there.